[expo-modules-core][android] refactor emulator detection
