+++ packages/expo-constants/android/src/main/java/expo/modules/constants/ConstantsService.kt
@@ -5,6 +5,7 @@ import org.apache.commons.io.IOUtils
 import com.facebook.device.yearclass.YearClass
 
 import expo.modules.core.interfaces.InternalModule
+import expo.modules.core.utilities.EmulatorUtilities
 import expo.modules.interfaces.constants.ConstantsInterface
 
 import android.os.Build
@@ -122,29 +123,7 @@ open class ConstantsService(private val context: Context) : InternalModule, Cons
     }
 
     private val isRunningOnEmulator: Boolean
-      get() = Build.FINGERPRINT.startsWith("generic") ||
-        Build.FINGERPRINT.startsWith("unknown") ||
-        Build.MODEL.contains("google_sdk") ||
-        Build.MODEL.lowercase(Locale.ROOT).contains("droid4x") ||
-        Build.MODEL.contains("Emulator") ||
-        Build.MODEL.contains("Android SDK built for x86") ||
-        Build.MANUFACTURER.contains("Genymotion") ||
-        Build.HARDWARE.contains("goldfish") ||
-        Build.HARDWARE.contains("ranchu") ||
-        Build.HARDWARE.contains("vbox86") ||
-        Build.PRODUCT.contains("sdk") ||
-        Build.PRODUCT.contains("google_sdk") ||
-        Build.PRODUCT.contains("sdk_google") ||
-        Build.PRODUCT.contains("sdk_x86") ||
-        Build.PRODUCT.contains("vbox86p") ||
-        Build.PRODUCT.contains("emulator") ||
-        Build.PRODUCT.contains("simulator") ||
-        Build.BOARD.lowercase(Locale.ROOT).contains("nox") ||
-        Build.BOOTLOADER.lowercase(Locale.ROOT).contains("nox") ||
-        Build.HARDWARE.lowercase(Locale.ROOT).contains("nox") ||
-        Build.PRODUCT.lowercase(Locale.ROOT).contains("nox") ||
-        Build.SERIAL.lowercase(Locale.ROOT).contains("nox") ||
-        (Build.BRAND.startsWith("generic") && Build.DEVICE.startsWith("generic"))
+      get() = EmulatorUtilities.isRunningOnEmulator()
 
     private fun getLongVersionCode(info: PackageInfo) =
       if (Build.VERSION.SDK_INT >= 28) info.longVersionCode
