+++ packages/expo-device/android/src/main/java/expo/modules/device/DeviceModule.kt
@@ -3,6 +3,7 @@ package expo.modules.device
 import expo.modules.core.ExportedModule
 import expo.modules.core.Promise
 import expo.modules.core.interfaces.ExpoMethod
+import expo.modules.core.utilities.EmulatorUtilities
 
 import com.facebook.device.yearclass.YearClass
 
@@ -152,29 +153,7 @@ class DeviceModule(private val mContext: Context) : ExportedModule(mContext) {
     private val TAG = DeviceModule::class.java.simpleName
 
     private val isRunningOnEmulator: Boolean
-      get() = Build.FINGERPRINT.startsWith("generic") ||
-        Build.FINGERPRINT.startsWith("unknown") ||
-        Build.MODEL.contains("google_sdk") ||
-        Build.MODEL.lowercase(Locale.ROOT).contains("droid4x") ||
-        Build.MODEL.contains("Emulator") ||
-        Build.MODEL.contains("Android SDK built for x86") ||
-        Build.MANUFACTURER.contains("Genymotion") ||
-        Build.HARDWARE.contains("goldfish") ||
-        Build.HARDWARE.contains("ranchu") ||
-        Build.HARDWARE.contains("vbox86") ||
-        Build.PRODUCT.contains("sdk") ||
-        Build.PRODUCT.contains("google_sdk") ||
-        Build.PRODUCT.contains("sdk_google") ||
-        Build.PRODUCT.contains("sdk_x86") ||
-        Build.PRODUCT.contains("vbox86p") ||
-        Build.PRODUCT.contains("emulator") ||
-        Build.PRODUCT.contains("simulator") ||
-        Build.BOARD.lowercase(Locale.ROOT).contains("nox") ||
-        Build.BOOTLOADER.lowercase(Locale.ROOT).contains("nox") ||
-        Build.HARDWARE.lowercase(Locale.ROOT).contains("nox") ||
-        Build.PRODUCT.lowercase(Locale.ROOT).contains("nox") ||
-        Build.SERIAL.lowercase(Locale.ROOT).contains("nox") ||
-        (Build.BRAND.startsWith("generic") && Build.DEVICE.startsWith("generic"))
+      get() = EmulatorUtilities.isRunningOnEmulator()
 
     private fun getDeviceType(context: Context): DeviceType {
       // Detect TVs via UI mode (Android TVs) or system features (Fire TV).
